Finished BallCollector.java for now...

+++ src/storm/interfaces/IBallCollector.java
@@ -5,7 +5,7 @@ public interface IBallCollector {
     public static final double MOTOR_SPEED_IN = 1.0d;
     public static final double MOTOR_SPEED_OUT = -1.0d;
     
-    public void startCollecting(double direction);
+    public void startCollecting(int direction);
     public void stopCollecting();
     public void returnControl();
